+++ src/App/Documentation/utils/index.js
@@ -1,17 +1,19 @@
-import ComponentPreview, { Attribute, PxScript } from "./ComponentPreview";
+import Attribute from "./Attribute";
+import ComponentPreview from "./ComponentPreview";
 import DocHeading from "./DocHeading";
 import DocToc from "./DocToc";
-import RenderRoutes from "./RenderRoutes";
-import Icon from "./Icon";
 import ExperimentalComponentAlert from "./ExperimentalComponentAlert";
+import Icon from "./Icon";
+import PxScript from "./PxScript";
+import RenderRoutes from "./RenderRoutes";
 
 export {
-    ComponentPreview,
     Attribute,
-    PxScript,
+    ComponentPreview,
     DocHeading,
     DocToc,
-    RenderRoutes,
+    ExperimentalComponentAlert,
     Icon,
-    ExperimentalComponentAlert
+    PxScript,
+    RenderRoutes
 };