+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Changed loader component to be visible by default, use `.d-none` when needed.
 - Changed modal close button to be wrapped by an anchor tag to help keyboard navigation.
 - Changed alert close button to be wrapped by an anchor tag to help keyboard navigation.
+- Slightly changed the documentation example and styling for desctiption-table.
 
 ## [0.8.1] - 2018-08-16